[web] Split database.js to individual files
