+++ Makefile
@@ -1,3 +1,6 @@
-.PHONY: dev
+.PHONY: dev check
 dev:
 	yarn run docs:dev
+
+check:
+	cd data && node check.js
