+++ data/check.js
@@ -1,0 +1,4 @@
+import { databases } from './database/index.js'
+
+// TODO: check if data are consistent, now we are just checking import
+console.log(databases)
