+++ data/database/heroic.js
@@ -1,0 +1,14 @@
+let heroic =  {
+  name: 'heroic',
+  displayName: 'Heroic',
+  status: 'maintained',
+  lang: 'java',
+  backend: [
+    'bigtable',
+    'cassandra',
+    'elasticsearch'
+  ],
+  license: 'apache-2.0'
+}
+
+export { heroic }
