+++ data/database/influxdb.js
@@ -1,0 +1,15 @@
+let influxdb = {
+  name: 'influxdb',
+  displayName: 'InfluxDB',
+  status: 'active',
+  lang: 'go',
+  backend: [
+    'localfs'
+  ],
+  protocol: [
+    'http'
+  ],
+  license: 'mit'
+}
+
+export { influxdb }
