+++ data/database/kairosdb.js
@@ -1,0 +1,15 @@
+let kairosdb = {
+  name: 'kairosdb',
+  displayName: 'KairosDB',
+  status: 'active',
+  lang: 'java',
+  backend: [
+    'cassandra'
+  ],
+  protocol: [
+    'http'
+  ],
+  license: 'apache-2.0'
+}
+
+export { kairosdb }
