+++ data/database/opentsdb.js
@@ -1,0 +1,12 @@
+let opentsdb = {
+  name: 'opentsdb',
+  displayName: 'OpenTSDB',
+  status: 'active',
+  lang: 'java',
+  backend: [
+    'hbase'
+  ],
+  license: 'lgpl'
+}
+
+export { opentsdb }
