+++ data/database/xephonk.js
@@ -1,0 +1,13 @@
+let xephonk =   {
+  name: 'xephonk',
+  displayName: 'XephonK',
+  status: 'dead',
+  lang: 'go',
+  backend: [
+    'cassandra',
+    'localfs'
+  ],
+  license: 'mit'
+}
+
+export { xephonk }
