+++ package.json
@@ -1,5 +1,6 @@
 {
   "license": "MIT",
+  "type": "module",
   "scripts": {
     "docs:dev": "vuepress dev docs",
     "docs:build": "vuepress build docs"
